Break CustomVizPluginsSettingsPage to multiple files
