+++ enterprise/frontend/src/metabase-enterprise/custom_viz/components/PluginListItem.tsx
@@ -1,0 +1,135 @@
+import { useCallback, useState } from "react";
+import { t } from "ttag";
+
+import {
+  useRefreshCustomVizPluginMutation,
+  useUpdateCustomVizPluginMutation,
+} from "metabase/api";
+import { Link } from "metabase/common/components/Link";
+import {
+  ActionIcon,
+  Box,
+  Flex,
+  Group,
+  Icon,
+  Menu,
+  Stack,
+  Text,
+} from "metabase/ui";
+import type { CustomVizPlugin, CustomVizPluginId } from "metabase-types/api";
+
+import S from "./CustomVizPluginsSettingsPage.module.css";
+import { PluginIconPreview } from "./PluginIconPreview";
+
+const BASE_PATH = "/admin/settings/custom-visualizations";
+
+type Props = {
+  plugin: CustomVizPlugin;
+  onDelete: (id: CustomVizPluginId) => void;
+};
+
+export function PluginListItem({ plugin, onDelete }: Props) {
+  const [isDeleting, setIsDeleting] = useState(false);
+  const [updatePlugin] = useUpdateCustomVizPluginMutation();
+  const [refreshPlugin, { isLoading: isRefreshing }] =
+    useRefreshCustomVizPluginMutation();
+
+  const handleDelete = useCallback(async () => {
+    setIsDeleting(true);
+    try {
+      await onDelete(plugin.id);
+    } finally {
+      setIsDeleting(false);
+    }
+  }, [plugin.id, onDelete]);
+
+  const handleToggleEnabled = useCallback(async () => {
+    await updatePlugin({ id: plugin.id, enabled: !plugin.enabled });
+  }, [plugin.id, plugin.enabled, updatePlugin]);
+
+  const handleRefresh = useCallback(async () => {
+    await refreshPlugin(plugin.id);
+  }, [plugin.id, refreshPlugin]);
+
+  return (
+    <Flex
+      component={Link}
+      to={`${BASE_PATH}/edit/${plugin.id}`}
+      justify="space-between"
+      align="center"
+      p="md"
+      className={S.pluginListItem}
+    >
+      <Group gap="md" align="center">
+        <PluginIconPreview plugin={plugin} />
+        <Stack gap={4}>
+          <Text fw={700}>{plugin.display_name}</Text>
+          <Group gap="xs">
+            <Text
+              component="a"
+              href={plugin.repo_url}
+              target="_blank"
+              rel="noopener noreferrer"
+              size="sm"
+              c="text-tertiary"
+              td="underline"
+              onClick={(e: React.MouseEvent) => e.stopPropagation()}
+            >
+              {plugin.repo_url}
+            </Text>
+            {plugin.resolved_commit && (
+              <>
+                <Text size="sm" c="text-tertiary">
+                  &bull;
+                </Text>
+                <Text size="sm" c="text-tertiary">
+                  {t`Commit`}: {plugin.resolved_commit.slice(0, 8)}
+                </Text>
+              </>
+            )}
+          </Group>
+          {plugin.error_message && (
+            <Text size="sm" c="error">
+              {plugin.error_message}
+            </Text>
+          )}
+        </Stack>
+      </Group>
+      <Box
+        onClick={(e: React.MouseEvent) => {
+          e.preventDefault();
+          e.stopPropagation();
+        }}
+      >
+        <Menu>
+          <Menu.Target>
+            <ActionIcon variant="subtle" loading={isRefreshing || isDeleting}>
+              <Icon name="ellipsis" />
+            </ActionIcon>
+          </Menu.Target>
+          <Menu.Dropdown>
+            <Menu.Item
+              leftSection={<Icon name="refresh" />}
+              onClick={handleRefresh}
+            >
+              {t`Re-fetch`}
+            </Menu.Item>
+            <Menu.Item
+              leftSection={<Icon name={plugin.enabled ? "stop" : "play"} />}
+              onClick={handleToggleEnabled}
+            >
+              {plugin.enabled ? t`Disable` : t`Enable`}
+            </Menu.Item>
+            <Menu.Item
+              leftSection={<Icon name="trash" />}
+              color="error"
+              onClick={handleDelete}
+            >
+              {t`Remove`}
+            </Menu.Item>
+          </Menu.Dropdown>
+        </Menu>
+      </Box>
+    </Flex>
+  );
+}
